Introduced some regression into feature/fruit after some api testing using postman. I have started unit testing for the book feature of the application.

+++ shop/src/main/java/com/burmau/shop/model/book/Book.java
@@ -1,6 +1,9 @@
 package com.burmau.shop.model.book;
 
+import java.math.BigDecimal;
+
 public abstract class Book {
     abstract String getType();
-    abstract double getPrice();
+    abstract BigDecimal getPrice();
+
 }
